Rewrites the cookbook to adhere to best practices.
/cc @someara

This is a complete rewrite of the Maven cookbook which brings it up to
best practices. It does bump a whole major release because of some
backwards incompatible changes such as deprecating the Maven 2.0 usage
since it is now EOL. There is also a change in name for the resource
which manages maven artifacts.