Remove Java installation from the cookbook
Maven 3 requires Java 8, which requires some thought to install.  The default java::default recipe will not leave you with a openjdk release that includes Java 8.  To get Java 8 on all platforms you need to use the Oracle JDK which requires agreeing to the license agreement, which we can't do for users.  Instead we need to rely on the user to install Java.