Add ability to install Maven under different user/group
In order to install Maven as another user, we will need to create a new
group and user (in that order, so the user can refer to the group).

Tests have been re-organised in order to show separation of concerns.

Ensure that we also have Windows support by specifying the default
username/group depending on platform type.

Closes #82.

Signed-off-by: Jamie Tanna <jamie@jamietanna.co.uk>